Revert "Revert "Refactor: Standardize user balance data type to float""

+++ app/src/main/java/com/example/mandelamoney/view/Iface/IHomeDashboardView.java
@@ -13,7 +13,7 @@ import java.util.List;
 public interface IHomeDashboardView {
     void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState);
 
-    void displayBalance(double balance);
+    void displayBalance(float balance);
 
     void displayUserName(String name);
